+++ shenyu-e2e/pom.xml
@@ -23,7 +23,7 @@
         <groupId>org.apache</groupId>
         <artifactId>apache</artifactId>
         <version>21</version>
-        <relativePath>./pom.xml</relativePath>
+        <relativePath/>
     </parent>
     <groupId>org.apache.shenyu</groupId>
     <artifactId>shenyu-e2e</artifactId>
@@ -41,7 +41,6 @@
 
     <properties>
         <java.version>8</java.version>
-        <testcontainers.version>1.18.3</testcontainers.version>
         <junit.version>5.8.2</junit.version>
         <assertj.version>3.22.0</assertj.version>
         <hamcrest.version>1.3</hamcrest.version>
@@ -54,9 +53,10 @@
         <rest-assured.version>5.2.0</rest-assured.version>
         <jetbrains-annotations.version>23.0.0</jetbrains-annotations.version>
         <apache-rat-plugin.version>0.15</apache-rat-plugin.version>
-        <spring-boot.version>2.7.13</spring-boot.version>
         <spring-framework.version>5.3.28</spring-framework.version>
         <maven-checkstyle-plugin.version>3.1.0</maven-checkstyle-plugin.version>
+        <guava.version>31.1-jre</guava.version>
+        <commons-collection.verion>4.4</commons-collection.verion>
     </properties>
 
     <modules>
@@ -179,14 +179,6 @@
                 <groupId>org.junit</groupId>
                 <artifactId>junit-bom</artifactId>
                 <version>${junit.version}</version>
-                <type>pom</type>
-                <scope>import</scope>
-            </dependency>
-
-            <dependency>
-                <groupId>org.testcontainers</groupId>
-                <artifactId>testcontainers-bom</artifactId>
-                <version>${testcontainers.version}</version>
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
