+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-http/src/test/java/org/apache/shenyu/e2e/testcase/divide/DividePluginTest.java
@@ -25,7 +25,7 @@ import org.apache.shenyu.e2e.engine.scenario.specification.AfterEachSpec;
 import org.apache.shenyu.e2e.engine.scenario.specification.BeforeEachSpec;
 import org.apache.shenyu.e2e.engine.scenario.specification.CaseSpec;
 import org.apache.shenyu.e2e.enums.ServiceTypeEnum;
-import org.apache.shenyu.e2e.k8s.engine.annotation.ShenYuE2ETest;
+import org.apache.shenyu.e2e.engine.annotation.ShenYuTest;
 import org.apache.shenyu.e2e.model.ResourcesData;
 import org.apache.shenyu.e2e.model.ResourcesData.Resource;
 import org.apache.shenyu.e2e.model.response.SelectorDTO;
@@ -36,42 +36,21 @@ import org.junit.jupiter.api.BeforeEach;
 
 import java.util.List;
 
-//@ShenYuTest(
-//        mode = Mode.DOCKER,
-//        services = {
-//                @ServiceConfigure(
-//                        serviceName = "admin",
-//                        port = 9095,
-//                        baseUrl = "http://{hostname:localhost}:9095",
-//                        parameters = {
-//                                @Parameter(key = "username", value = "admin"),
-//                                @Parameter(key = "password", value = "123456")
-//                        }
-//                ),
-//                @ServiceConfigure(
-//                        serviceName = "gateway",
-//                        port = 9195,
-//                        baseUrl = "http://{hostname:localhost}:9195",
-//                        type = ServiceType.SHENYU_GATEWAY
-//                )
-//        },
-//        dockerComposeFile = "classpath:./docker-compose.{storage:h2}.yml"
-//)
-@ShenYuE2ETest(environments = {
-        @ShenYuE2ETest.Environment(
+@ShenYuTest(environments = {
+        @ShenYuTest.Environment(
                 serviceName = "shenyu-e2e-admin",
-                service = @ShenYuE2ETest.ServiceConfigure(moduleName = "shenyu-e2e",
+                service = @ShenYuTest.ServiceConfigure(moduleName = "shenyu-e2e",
                         baseUrl = "http://localhost:31095",
                         type = ServiceTypeEnum.SHENYU_ADMIN,
                         parameters = {
-                                @ShenYuE2ETest.Parameter(key = "username", value = "admin"),
-                                @ShenYuE2ETest.Parameter(key = "password", value = "123456")
+                                @ShenYuTest.Parameter(key = "username", value = "admin"),
+                                @ShenYuTest.Parameter(key = "password", value = "123456")
                         }
                 )
         ),
-        @ShenYuE2ETest.Environment(
+        @ShenYuTest.Environment(
                 serviceName = "shenyu-e2e-gateway",
-                service = @ShenYuE2ETest.ServiceConfigure(moduleName = "shenyu-e2e",
+                service = @ShenYuTest.ServiceConfigure(moduleName = "shenyu-e2e",
                         baseUrl = "http://localhost:31195",
                         type = ServiceTypeEnum.SHENYU_GATEWAY
                 )
