+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-spring-cloud/src/test/java/org/apache/shenyu/e2e/testcase/springcloud/DataSynTest.java
@@ -21,7 +21,7 @@ import org.apache.shenyu.e2e.client.WaitDataSync;
 import org.apache.shenyu.e2e.client.admin.AdminClient;
 import org.apache.shenyu.e2e.client.gateway.GatewayClient;
 import org.apache.shenyu.e2e.enums.ServiceTypeEnum;
-import org.apache.shenyu.e2e.k8s.engine.annotation.ShenYuE2ETest;
+import org.apache.shenyu.e2e.engine.annotation.ShenYuTest;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -53,21 +53,21 @@ import org.junit.jupiter.api.Test;
 //        },
 //        dockerComposeFile = "classpath:./docker-compose.mysql.yml"
 //)
-@ShenYuE2ETest(environments = {
-        @ShenYuE2ETest.Environment(
+@ShenYuTest(environments = {
+        @ShenYuTest.Environment(
                 serviceName = "shenyu-e2e-admin",
-                service = @ShenYuE2ETest.ServiceConfigure(moduleName = "shenyu-e2e",
+                service = @ShenYuTest.ServiceConfigure(moduleName = "shenyu-e2e",
                         baseUrl = "http://localhost:31095",
                         type = ServiceTypeEnum.SHENYU_ADMIN,
                         parameters = {
-                                @ShenYuE2ETest.Parameter(key = "username", value = "admin"),
-                                @ShenYuE2ETest.Parameter(key = "password", value = "123456")
+                                @ShenYuTest.Parameter(key = "username", value = "admin"),
+                                @ShenYuTest.Parameter(key = "password", value = "123456")
                         }
                 )
         ),
-        @ShenYuE2ETest.Environment(
+        @ShenYuTest.Environment(
                 serviceName = "shenyu-e2e-gateway",
-                service = @ShenYuE2ETest.ServiceConfigure(moduleName = "shenyu-e2e",
+                service = @ShenYuTest.ServiceConfigure(moduleName = "shenyu-e2e",
                         baseUrl = "http://localhost:31195",
                         type = ServiceTypeEnum.SHENYU_GATEWAY
                 )
