+++ shenyu-e2e/shenyu-e2e-case/shenyu-e2e-case-spring-cloud/src/test/java/org/apache/shenyu/e2e/testcase/springcloud/SpringCloudPluginTest.java
@@ -17,19 +17,19 @@
 
 package org.apache.shenyu.e2e.testcase.springcloud;
 
+import com.google.common.collect.Lists;
 import org.apache.shenyu.e2e.client.WaitDataSync;
 import org.apache.shenyu.e2e.client.admin.AdminClient;
 import org.apache.shenyu.e2e.client.gateway.GatewayClient;
 import org.apache.shenyu.e2e.engine.annotation.ShenYuScenario;
 import org.apache.shenyu.e2e.engine.scenario.specification.CaseSpec;
 import org.apache.shenyu.e2e.enums.ServiceTypeEnum;
-import org.apache.shenyu.e2e.k8s.engine.annotation.ShenYuE2ETest;
+import org.apache.shenyu.e2e.engine.annotation.ShenYuTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
-import org.testcontainers.shaded.com.google.common.collect.Lists;
 
 import java.util.List;
 
@@ -53,21 +53,21 @@ import java.util.List;
 //                        serviceName = "gateway",
 //                        port = 9195,
 //                        baseUrl = "http://{hostname:localhost}:9195", type = ShenYuEngineConfigure.ServiceType.SHENYU_GATEWAY, parameters = {@ShenYuTest.Parameter(key = "application", value = "spring.cloud.discovery.enabled:true,eureka.client.enabled:true"), @ShenYuTest.Parameter(key = "dataSyn", value = "gateway_websocket")})}, dockerComposeFile = "classpath:./docker-compose.mysql.yml")
-@ShenYuE2ETest(environments = {
-        @ShenYuE2ETest.Environment(
+@ShenYuTest(environments = {
+        @ShenYuTest.Environment(
                 serviceName = "shenyu-e2e-admin",
-                service = @ShenYuE2ETest.ServiceConfigure(moduleName = "shenyu-e2e",
+                service = @ShenYuTest.ServiceConfigure(moduleName = "shenyu-e2e",
                         baseUrl = "http://localhost:31095",
                         type = ServiceTypeEnum.SHENYU_ADMIN,
                         parameters = {
-                                @ShenYuE2ETest.Parameter(key = "username", value = "admin"),
-                                @ShenYuE2ETest.Parameter(key = "password", value = "123456")
+                                @ShenYuTest.Parameter(key = "username", value = "admin"),
+                                @ShenYuTest.Parameter(key = "password", value = "123456")
                         }
                 )
         ),
-        @ShenYuE2ETest.Environment(
+        @ShenYuTest.Environment(
                 serviceName = "shenyu-e2e-gateway",
-                service = @ShenYuE2ETest.ServiceConfigure(moduleName = "shenyu-e2e",
+                service = @ShenYuTest.ServiceConfigure(moduleName = "shenyu-e2e",
                         baseUrl = "http://localhost:31195",
                         type = ServiceTypeEnum.SHENYU_GATEWAY
                 )
