+++ shenyu-e2e/shenyu-e2e-common/pom.xml
@@ -44,13 +44,7 @@
         <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-collections4</artifactId>
-            <version>4.4</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>31.1-jre</version>
+            <version>${commons-collection.verion}</version>
         </dependency>
     </dependencies>
