+++ shenyu-e2e/shenyu-e2e-engine/src/main/java/org/apache/shenyu/e2e/engine/ShenYuExtensionContext.java
@@ -17,103 +17,101 @@
 
 package org.apache.shenyu.e2e.engine;
 
-import junit.framework.AssertionFailedError;
+import org.apache.commons.collections4.MapUtils;
+import org.apache.shenyu.e2e.client.EnvironmentClient;
 import org.apache.shenyu.e2e.client.ExternalServiceClient;
 import org.apache.shenyu.e2e.client.admin.AdminClient;
 import org.apache.shenyu.e2e.client.gateway.GatewayClient;
 import org.apache.shenyu.e2e.common.IdGenerator;
+import org.apache.shenyu.e2e.config.ServiceConfigure;
+import org.apache.shenyu.e2e.enums.ServiceTypeEnum;
 import org.apache.shenyu.e2e.engine.config.ShenYuEngineConfigure;
-import org.apache.shenyu.e2e.engine.service.DockerServiceCompose;
-import org.apache.shenyu.e2e.engine.service.HostServiceCompose;
-import org.apache.shenyu.e2e.engine.service.ServiceCompose;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
 
-import static org.apache.shenyu.e2e.engine.config.ShenYuEngineConfigure.Mode.DOCKER;
-import static org.apache.shenyu.e2e.engine.config.ShenYuEngineConfigure.Mode.HOST;
-
-/**
- * Start the project using the corresponding method based on the selected configuration.
- */
 public class ShenYuExtensionContext {
 
     private final String scenarioId;
+    
+    private final EnvironmentClient environmentClient = new EnvironmentClient();
 
-    private final ServiceCompose serviceCompose;
+    private Map<String, AdminClient> adminClientMap;
 
-    private AdminClient adminClient;
-
-    private GatewayClient gatewayClient;
+    private Map<String, GatewayClient> gatewayClientMap;
 
     private Map<String, ExternalServiceClient> externalServiceClientMap;
 
-    ShenYuExtensionContext(final ShenYuEngineConfigure config) {
-        if (config.getMode() == HOST) {
-            serviceCompose = new HostServiceCompose(config.getHostConfigure());
-        } else if (config.getMode() == DOCKER) {
-            serviceCompose = new DockerServiceCompose(config.getDockerConfigure());
-        } else {
-            throw new AssertionFailedError("Mode [" + config.getMode() + "] is not supported yet");
-        }
+    /**
+     * Instantiates a new Shenyu e2e extension context.
+     */
+    public ShenYuExtensionContext(final ShenYuEngineConfigure configure) {
         this.scenarioId = IdGenerator.generateScenarioId();
+        Map<String, ServiceConfigure> serviceConfigureMap = configure.getServiceConfigureMap();
+        serviceConfigureMap.forEach((serviceName, serviceConfigure) -> {
+            if (ServiceTypeEnum.SHENYU_ADMIN.equals(serviceConfigure.getServiceType())) {
+                if (MapUtils.isEmpty(adminClientMap)) {
+                    adminClientMap = new HashMap<>();
+                }
+                AdminClient client = new AdminClient(scenarioId, serviceName, serviceConfigure.getBaseUrl(), serviceConfigure.getParameters());
+                adminClientMap.put(serviceName, client);
+                environmentClient.add(client);
+            } else if (ServiceTypeEnum.SHENYU_GATEWAY.equals(serviceConfigure.getServiceType())) {
+                if (MapUtils.isEmpty(gatewayClientMap)) {
+                    gatewayClientMap = new HashMap<>();
+                }
+                GatewayClient client = new GatewayClient(scenarioId, serviceName, serviceConfigure.getBaseUrl(), serviceConfigure.getParameters());
+                gatewayClientMap.put(serviceName, client);
+                environmentClient.add(client);
+            } else if (ServiceTypeEnum.EXTERNAL_SERVICE.equals(serviceConfigure.getServiceType())) {
+                if (MapUtils.isEmpty(externalServiceClientMap)) {
+                    externalServiceClientMap = new HashMap<>();
+                }
+                ExternalServiceClient client = new ExternalServiceClient(scenarioId, serviceName,
+                        serviceConfigure.getBaseUrl(), serviceConfigure.getParameters());
+                externalServiceClientMap.put(serviceName, client);
+                environmentClient.add(client);
+            }
+        });
+        
     }
 
     /**
-     * start.
-     */
-    public void setup() {
-        serviceCompose.start();
-    }
-
-    /**
-     * get admin client.
-     * @return AdminClient
-     */
-    public AdminClient getAdminClient() {
-        if (Objects.isNull(adminClient)) {
-            adminClient = serviceCompose.newAdminClient(scenarioId);
-        }
-        return adminClient;
-    }
-
-    /**
-     * get gateway client.
-     * @return GatewayClient
-     */
-    public GatewayClient getGatewayClient() {
-        if (Objects.isNull(gatewayClient)) {
-            gatewayClient = serviceCompose.newGatewayClient(scenarioId);
-        }
-        return gatewayClient;
-    }
-
-    /**
-     * get external service client.
-     * @param externalServiceName externalServiceName
-     * @return ExternalServiceClient
-     */
-    public ExternalServiceClient getExternalServiceClient(final String externalServiceName) {
-        if (Objects.isNull(externalServiceClientMap)) {
-            externalServiceClientMap = new HashMap<>();
-        }
-        return externalServiceClientMap.computeIfAbsent(externalServiceName, serviceCompose::newExternalServiceClient);
-    }
-
-    /**
-     * clean up.
-     */
-    public void cleanup() {
-        serviceCompose.stop();
-    }
-
-    /**
-     * get scenarioId.
+     * Gets admin client map.
      *
-     * @return scenarioId
+     * @return current admin client map
      */
-    public String getScenarioId() {
-        return scenarioId;
+    public Map<String, AdminClient> getAdminClientMap() {
+        return adminClientMap;
+    }
+
+    /**
+     * Gets gateway client map.
+     *
+     * @return current gateway client map
+     */
+    public Map<String, GatewayClient> getGatewayClientMap() {
+        return gatewayClientMap;
+    }
+    
+
+
+    /**
+     * Gets external service client map.
+     *
+     * @return current external service client map
+     */
+    public Map<String, ExternalServiceClient> getExternalServiceClientMap() {
+        return externalServiceClientMap;
+    }
+
+    
+    /**
+     * Gets environment client.
+     *
+     * @return current environment client
+     */
+    public EnvironmentClient getEnvironmentClient() {
+        return environmentClient;
     }
 }
