--- shenyu-e2e/shenyu-e2e-engine/src/main/java/org/apache/shenyu/e2e/k8s/engine/ShenYuLogExtension.java
@@ -1,61 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.e2e.k8s.engine;
-
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.junit.jupiter.api.extension.InvocationInterceptor;
-import org.junit.jupiter.api.extension.ReflectiveInvocationContext;
-import org.slf4j.MDC;
-
-import java.lang.reflect.Method;
-
-/**
- * ShenYu log extension.
- */
-public class ShenYuLogExtension implements InvocationInterceptor {
-    
-    @Override
-    public void interceptBeforeAllMethod(final Invocation<Void> invocation, final ReflectiveInvocationContext<Method> invocationContext, final ExtensionContext extensionContext) throws Throwable {
-        MDC.put("phase", "BeforeAll");
-        invocation.proceed();
-    }
-    
-    @Override
-    public void interceptBeforeEachMethod(final Invocation<Void> invocation, final ReflectiveInvocationContext<Method> invocationContext, final ExtensionContext extensionContext) throws Throwable {
-        MDC.put("phase", "BeforeEach");
-        invocation.proceed();
-    }
-    
-    @Override
-    public void interceptTestTemplateMethod(final Invocation<Void> invocation, final ReflectiveInvocationContext<Method> invocationContext, final ExtensionContext extensionContext) throws Throwable {
-        MDC.put("phase", "Verify");
-        invocation.proceed();
-    }
-    
-    @Override
-    public void interceptAfterEachMethod(final Invocation<Void> invocation, final ReflectiveInvocationContext<Method> invocationContext, final ExtensionContext extensionContext) throws Throwable {
-        MDC.put("phase", "AfterEach");
-        invocation.proceed();
-    }
-    
-    @Override
-    public void interceptAfterAllMethod(final Invocation<Void> invocation, final ReflectiveInvocationContext<Method> invocationContext, final ExtensionContext extensionContext) throws Throwable {
-        MDC.put("phase", "AfterAll");
-        invocation.proceed();
-    }
-}
